+++ compiler-tests/src/test/java/dev/zacsweers/metro/compiler/DiagnosticTestGenerated.java
@@ -459,6 +459,12 @@ public class DiagnosticTestGenerated extends AbstractDiagnosticTest {
     }
 
     @Test
+    @TestMetadata("DependencyGraphFactory_ParamsCannotBeScoped.kt")
+    public void testDependencyGraphFactory_ParamsCannotBeScoped() {
+      run("DependencyGraphFactory_ParamsCannotBeScoped.kt");
+    }
+
+    @Test
     @TestMetadata("DependencyGraphFactory_ProvidesParamCannotBeIntrinsic.kt")
     public void testDependencyGraphFactory_ProvidesParamCannotBeIntrinsic() {
       run("DependencyGraphFactory_ProvidesParamCannotBeIntrinsic.kt");
@@ -621,6 +627,12 @@ public class DiagnosticTestGenerated extends AbstractDiagnosticTest {
       @TestMetadata("ExtensionsWithFactoriesMustUseThem.kt")
       public void testExtensionsWithFactoriesMustUseThem() {
         run("ExtensionsWithFactoriesMustUseThem.kt");
+      }
+
+      @Test
+      @TestMetadata("GraphExtensionFactory_ParamsCannotBeScoped.kt")
+      public void testGraphExtensionFactory_ParamsCannotBeScoped() {
+        run("GraphExtensionFactory_ParamsCannotBeScoped.kt");
       }
 
       @Test